final changes for pmax 4.4-Lite II release
SCCS-vsn: 8.5

+++ sys/pmax/dev/tz.c
@@ -33,7 +33,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)tz.c	8.4 (Berkeley) 01/11/94
+ *	@(#)tz.c	8.5 (Berkeley) 06/02/95
  *
  * from: $Header: /sprite/src/kernel/dev/RCS/devSCSITape.c,
  *	v 8.14 89/07/31 17:26:13 mendel Exp $ SPRITE (Berkeley)
@@ -696,7 +696,7 @@ tzclose(dev, flag)
 
 tzioctl(dev, cmd, data, flag)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 {
